services files
One is for running tests directly in selected folder. Currently is for py.
One is for copying all folder's files to another location and installing dependencies necessary then running tests in that folder. Currently is for js so that I can store and edit source files on any cloud drive and don't have to store dependencies on cloud which are too many even only for one Tony project.